+++ src/conf/mod.rs
@@ -5,11 +5,13 @@ use {
 
 mod conf;
 mod format;
+mod import;
 mod verb_conf;
 
 pub use {
     conf::Conf,
     format::*,
+    import::*,
     once_cell::sync::Lazy,
     verb_conf::VerbConf,
 };